rename universalheader to headserach to  match function name of php file
