+++ src/functions.php
@@ -51,7 +51,7 @@ function utkchancellor_sidebar() {
 
 require_once ( 'functions/filter-mobile-nav.php' );
 require_once ( 'functions/filter-region-footer.php' );
-require_once ( 'functions/filter-region-universalheader.php' );
+require_once ( 'functions/filter-region-headsearch.php' );
 
 
 // Functions to modify elements
